move user building to settings
