+++ src/main/java/com/plux/distribution/Main.java
@@ -10,17 +10,17 @@ import com.plux.distribution.core.workflow.application.frame.registration.hello.
 import com.plux.distribution.core.workflow.application.frame.registration.pin.CheckPasswordFrame;
 import com.plux.distribution.core.workflow.application.frame.registration.pin.CorrectPasswordFrame;
 import com.plux.distribution.core.workflow.application.frame.registration.pin.InorrectPasswordFrame;
-import com.plux.distribution.core.workflow.application.frame.registration.user.AskAgeFrame;
-import com.plux.distribution.core.workflow.application.frame.registration.user.AskCityFrame;
-import com.plux.distribution.core.workflow.application.frame.registration.user.AskEmailFrame;
-import com.plux.distribution.core.workflow.application.frame.registration.user.AskHobbyFrame;
-import com.plux.distribution.core.workflow.application.frame.registration.user.AskNameFrame;
+import com.plux.distribution.core.workflow.application.frame.settings.user.AskAgeFrame;
+import com.plux.distribution.core.workflow.application.frame.settings.user.AskCityFrame;
+import com.plux.distribution.core.workflow.application.frame.settings.user.AskEmailFrame;
+import com.plux.distribution.core.workflow.application.frame.settings.user.AskHobbyFrame;
+import com.plux.distribution.core.workflow.application.frame.settings.user.AskNameFrame;
 import com.plux.distribution.core.workflow.application.frame.settings.SettingsAppliedMessage;
 import com.plux.distribution.core.workflow.application.frame.settings.schedule.AskHoursFrame;
 import com.plux.distribution.core.workflow.application.frame.settings.schedule.AskSpdFrame;
 import com.plux.distribution.core.workflow.application.frame.settings.schedule.AskTimezoneFrame;
-import com.plux.distribution.core.workflow.application.frame.registration.user.FinalizeFrame;
-import com.plux.distribution.core.workflow.application.frame.registration.user.StartUserBuildingFrame;
+import com.plux.distribution.core.workflow.application.frame.settings.user.FinalizeFrame;
+import com.plux.distribution.core.workflow.application.frame.settings.user.StartUserBuildingFrame;
 import com.plux.distribution.core.workflow.application.frame.settings.schedule.FinalizeScheduleSettingsFrame;
 import com.plux.distribution.core.workflow.application.frame.settings.schedule.ScheduleSettingsBuilder;
 import com.plux.distribution.core.workflow.application.frame.settings.schedule.StartScheduleSettingsFrame;
@@ -42,7 +42,7 @@ import com.plux.distribution.core.workflow.application.port.out.DataRegistry;
 import com.plux.distribution.core.workflow.application.port.out.FrameRegistry;
 import com.plux.distribution.core.workflow.application.frame.DefaultDataRegistry;
 import com.plux.distribution.core.workflow.application.frame.DefaultFrameRegistry;
-import com.plux.distribution.core.workflow.application.frame.registration.user.UserBuilder;
+import com.plux.distribution.core.workflow.application.frame.settings.user.UserBuilder;
 import com.plux.distribution.core.workflow.application.frame.utils.LastMessageData;
 import com.plux.distribution.core.workflow.application.frame.utils.SequenceFrame;
 import com.plux.distribution.core.workflow.application.service.WorkflowService;
