+++ src/main/java/com/plux/distribution/core/workflow/application/frame/settings/user/AskEmailFrame.java
@@ -1,8 +1,9 @@
-package com.plux.distribution.core.workflow.application.frame.registration.user;
+package com.plux.distribution.core.workflow.application.frame.settings.user;
 
 import com.plux.distribution.core.workflow.domain.Frame;
 import com.plux.distribution.core.workflow.domain.FrameContext;
 import com.plux.distribution.core.workflow.domain.FrameFeedback;
+import com.plux.distribution.core.workflow.application.frame.utils.InfoMessageFrame;
 import com.plux.distribution.core.workflow.application.frame.utils.LastMessageData;
 import com.plux.distribution.core.message.application.dto.action.ClearButtonsAction;
 import com.plux.distribution.core.message.domain.attachment.ButtonAttachment;
@@ -10,12 +11,12 @@ import com.plux.distribution.core.message.domain.content.SimpleMessageContent;
 import java.util.List;
 import org.jetbrains.annotations.NotNull;
 
-public class AskHobbyFrame implements Frame {
+public class AskEmailFrame implements Frame {
 
     @Override
     public void exec(@NotNull FrameContext context) {
         var messageId = context.send(new SimpleMessageContent(
-                context.getTextProvider().getString("registration.user.hobby.ask"),
+                context.getTextProvider().getString("registration.user.email.ask"),
                 List.of(new ButtonAttachment(context.getTextProvider().getString("utils.skip_button"), "skip"))
         ));
 
@@ -28,14 +29,24 @@ public class AskHobbyFrame implements Frame {
 
         feedback.buttonTag().ifPresent(value -> {
             if (value.equals("skip")) {
-                userBuilder.setHobby(null);
+                userBuilder.setEmail(null);
                 goNext(context);
             }
         });
 
         feedback.text().ifPresent(text -> {
-            userBuilder.setHobby(text);
-            goNext(context);
+            try {
+                userBuilder.setEmail(text);
+                goNext(context);
+            } catch (IllegalArgumentException e) {
+                context.changeState(this, false);
+                context.push(
+                        new InfoMessageFrame(
+                                context.getTextProvider().getString("registration.user.email.wrong_type")
+                        ),
+                        true);
+                context.exec();
+            }
         });
     }
