+++ src/main/java/com/plux/distribution/core/workflow/application/frame/settings/user/AskHobbyFrame.java
@@ -1,9 +1,8 @@
-package com.plux.distribution.core.workflow.application.frame.registration.user;
+package com.plux.distribution.core.workflow.application.frame.settings.user;
 
 import com.plux.distribution.core.workflow.domain.Frame;
 import com.plux.distribution.core.workflow.domain.FrameContext;
 import com.plux.distribution.core.workflow.domain.FrameFeedback;
-import com.plux.distribution.core.workflow.application.frame.utils.InfoMessageFrame;
 import com.plux.distribution.core.workflow.application.frame.utils.LastMessageData;
 import com.plux.distribution.core.message.application.dto.action.ClearButtonsAction;
 import com.plux.distribution.core.message.domain.attachment.ButtonAttachment;
@@ -11,12 +10,12 @@ import com.plux.distribution.core.message.domain.content.SimpleMessageContent;
 import java.util.List;
 import org.jetbrains.annotations.NotNull;
 
-public class AskEmailFrame implements Frame {
+public class AskHobbyFrame implements Frame {
 
     @Override
     public void exec(@NotNull FrameContext context) {
         var messageId = context.send(new SimpleMessageContent(
-                context.getTextProvider().getString("registration.user.email.ask"),
+                context.getTextProvider().getString("registration.user.hobby.ask"),
                 List.of(new ButtonAttachment(context.getTextProvider().getString("utils.skip_button"), "skip"))
         ));
 
@@ -29,24 +28,14 @@ public class AskEmailFrame implements Frame {
 
         feedback.buttonTag().ifPresent(value -> {
             if (value.equals("skip")) {
-                userBuilder.setEmail(null);
+                userBuilder.setHobby(null);
                 goNext(context);
             }
         });
 
         feedback.text().ifPresent(text -> {
-            try {
-                userBuilder.setEmail(text);
-                goNext(context);
-            } catch (IllegalArgumentException e) {
-                context.changeState(this, false);
-                context.push(
-                        new InfoMessageFrame(
-                                context.getTextProvider().getString("registration.user.email.wrong_type")
-                        ),
-                        true);
-                context.exec();
-            }
+            userBuilder.setHobby(text);
+            goNext(context);
         });
     }
