+++ src/main/java/com/plux/distribution/core/workflow/application/frame/settings/user/UserBuilder.java
@@ -1,4 +1,4 @@
-package com.plux.distribution.core.workflow.application.frame.registration.user;
+package com.plux.distribution.core.workflow.application.frame.settings.user;
 
 import com.plux.distribution.core.workflow.application.utils.JsonDataSerializer;
 import com.plux.distribution.core.user.domain.UserInfo;
